feat(R4) Refactoring classes moved to utils
