+++ growingnn/structure.py
@@ -11,6 +11,11 @@ from .config import config, DistributionMode
 from .optimizers import *
 from .quaziIdentity import *
 from .utils.stoppers import BaseStopper, AccuracyStopper, ParameterCountStopper, AccuracyAndReductionStopper
+from .utils.loss import Loss
+from .utils.activations import Activations
+from .utils.lr_scheduler import LearningRateScheduler
+from .utils.history import History
+from .utils.storage import Storage
 
 def switch_to_gpu():
     global np, IS_CUPY, correlate, convolve
@@ -26,176 +31,9 @@ def switch_to_cpu():
     from scipy.signal import correlate2d
     from scipy.signal import convolve2d
 
-class Loss:
-    
-    def getByName(name):
-        if name == Loss.MSE.__name__:
-            return Loss.MSE
-        elif name == Loss.MAE.__name__:
-            return Loss.MAE
-        elif name == Loss.multiclass_cross_entropy.__name__:
-            return Loss.multiclass_cross_entropy
-
-    class MSE:
-        __name__ = 'MSE'
-        def exe(Y_true, Y_pred):
-            return np.sum((Y_pred - Y_true)**2)/Y_pred.shape[0]
-        def der(Y_true, Y_pred):
-            return Y_pred - Y_true
-    
-    class MAE:
-        __name__ = 'MAE'
-        def exe(Y_true, Y_pred):
-            return np.sum(np.abs(Y_pred - Y_true))/Y_pred.shape[0]
-        def der(Y_true, Y_pred):
-            return np.sign(Y_pred - Y_true)
-    
-    class multiclass_cross_entropy:
-        __name__ = 'multiclass_cross_entropy'
-        def exe(Y_true, Y_pred):
-            error = 0.0
-            for i in range(0, Y_true.shape[1]):
-                error -= np.dot(Y_true[:,i].T, np.log(Y_pred[:,i]))
-            #print("Y_true.shape: ", Y_true.shape)
-            return error / Y_true.shape[1]
-        def der(Y_true, Y_pred):
-            grad = np.zeros(Y_true.shape)
-            for i in range(0, Y_true.shape[1]):
-                #print("Y_true[:, i]: ", Y_true[:, i].shape)
-                #print("Y_pred[:, i]: ", Y_pred[:, i].shape)
-                partial_grad = -Y_true[:, i] / Y_pred[:, i]
-                A = np.tile(np.reshape(Y_pred[:, i], (Y_pred.shape[0], 1)), (1, Y_pred.shape[0]))
-                grad[:, i] = (A * (np.identity(Y_pred.shape[0]) - A.T)) @ partial_grad
-            return grad
-
-class Activations:
-
-    def getByName(name):
-        if name == Activations.ReLu.__name__:
-            return Activations.ReLu
-        elif name == Activations.leaky_ReLu.__name__:
-            return Activations.leaky_ReLu
-        elif name == Activations.SoftMax.__name__:
-            return Activations.SoftMax
-        elif name == Activations.Sigmoid.__name__:
-            return Activations.Sigmoid
-        elif name == Activations.Tanh.__name__:
-            return Activations.Tanh
-        elif name == Activations.Linear.__name__:
-            return Activations.Linear
-    
-    class ReLu:
-        __name__ = 'ReLu'
-
-        @staticmethod
-        @jit(nopython=True)
-        def exe(X):
-            return np.maximum(X,0)
-        
-        @staticmethod
-        @jit(nopython=True)
-        def der(X):
-            return np.where(X > 0, 1, 0)
-        
-    class leaky_ReLu:
-        __name__ = 'leaky_ReLu'
-
-        @staticmethod
-        @jit(nopython=True)
-        def exe(X):
-            return  np.where(X > 0, X, X * 0.001)
-        
-        @staticmethod
-        @jit(nopython=True)
-        def der(X):
-            return  np.where(X > 0, 1, 0.001)
-        
-    class SoftMax:
-        __name__ = 'SoftMax'
-
-        @staticmethod
-        def exe(X):
-            # Vectorized implementation instead of loop
-            exp_X = np.exp(X - np.max(X, axis=0))
-            result = exp_X / np.sum(exp_X, axis=0)
-            if config.ENABLE_CLIP_ON_ACTIVATIONS:
-                return clip(result, 0.0001, 0.999)
-            else:
-                return result
-        
-        @staticmethod
-        @jit(nopython=True)
-        def der(X):
-            return 1.0
-        
-    class Sigmoid:
-        __name__ = 'Sigmoid'
-
-        @staticmethod
-        @jit(nopython=True)
-        def exe(X):
-            return  1/(1 + np.exp(-X))
-        
-        @staticmethod
-        @jit(nopython=True)
-        def der(X):
-            sigm = 1/(1 + np.exp(-X))
-            return sigm * (1.0 - sigm)
-        
-    class Tanh:
-        __name__ = 'Tanh'
-
-        @staticmethod
-        @jit(nopython=True)
-        def exe(X):
-            return np.tanh(X)
-        
-        @staticmethod
-        @jit(nopython=True)
-        def der(X):
-            return 1 - np.tanh(X)**2
-    
-    class Linear:
-        __name__ = 'Linear'
-        
-        @staticmethod
-        @jit(nopython=True)
-        def exe(X):
-            return X
-        
-        @staticmethod
-        @jit(nopython=True)
-        def der(X):
-            return np.ones_like(X)
 
 
-class LearningRateScheduler:
-    CONSTANT = 0
-    PROGRESIVE = 1
-    PROGRESIVE_PARABOIDAL = 2
 
-    def __init__(self, mode, alpha, steepness = 0.2):
-        self.mode = mode
-        self.alpha = alpha
-        self.steepness = steepness
-
-    def alpha_scheduler(self, i, iterations):
-        thresh = float(self.steepness * iterations)
-        i = float(i)
-        iterations = float(iterations)
-        result = self.alpha
-        if self.mode == LearningRateScheduler.CONSTANT:
-            result =  self.alpha
-        elif self.mode == LearningRateScheduler.PROGRESIVE:
-            if i < thresh:
-                return self.alpha * ((i+1) / (thresh + 2))
-            result = self.alpha * (1 - (i - thresh) / (iterations - thresh + 2))
-        else:
-            thresh = self.steepness * iterations
-            if i < thresh:
-                return self.alpha * ( -1 * ((1) / (pow(thresh, 2))) * pow((i) - thresh, 2) + 1)
-            result = self.alpha * (-1 * ((1) / (pow(iterations - thresh, 2))) * pow((i) - thresh, 2) + 1)
-        return max(0, result)
 
 class SimulationScheduler:
     CONSTANT = 0
@@ -247,134 +85,6 @@ class SimulationScheduler:
         elif self.mode == SimulationScheduler.OVERFIT_CHECK: return "simoverfit"
 
 
-class History:
-    def __init__(self, keys):
-        self.Y = {}
-        self.last_img_id = 0
-        self.description = "\ndescription_of_training_process: \n"
-        self.best_train_acc = 0.0
-        self.best_test_acc = 0.0
-        for key in keys:
-            self.Y[key] = []
-
-    @staticmethod
-    @jit(nopython=True)
-    def _calculate_accuracy(correct_predictions, total_samples):
-        """Calculate accuracy using Numba JIT"""
-        return correct_predictions / total_samples
-
-    def update_training_progress(self, correct_predictions, total_samples, total_loss, epoch, current_alpha, quiet):
-        """Update training history and print progress"""
-        acc = self._calculate_accuracy(correct_predictions, total_samples)
-        self.append('accuracy', acc)
-        self.append('loss', total_loss)
-        return acc
-
-    def get_length(self):
-        return len(self.Y[list(self.Y.keys())[0]])
-
-    def append(self, key, value):
-        if not key in self.Y.keys():
-            self.Y[key] = []
-        self.Y[key].append(value)
-
-    def merge(self, new_hist):
-        for key in self.Y.keys():
-            if key in self.Y.keys() and key in new_hist.Y.keys():
-                self.Y[key] += new_hist.Y[key]
-
-    def _check_learning_capability(accuracies, patience=15, verbose=0.3): #0.425
-        """
-        Determine if a model is still learning based on recent accuracy values.
-        Returns True if the model is still learning (significant upward trend), 
-        or False if learning has plateaued or stalled.
-        """
-        import numpy as np
-        
-        # If not enough data points, assume the model can still learn
-        if len(accuracies) < 2:
-            return True
-        
-        # Consider only the most recent `patience` accuracy values
-        recent_values = accuracies[-patience:] if patience > 0 else accuracies
-        if len(recent_values) < 2:
-            # Not enough values in the window to determine a trend
-            return True
-        
-        # Calculate the overall slope as the difference between the last and first accuracy in the window
-        net_improvement = recent_values[-1] - recent_values[0]
-        # Calculate the standard deviation of accuracies in the window to gauge fluctuations
-        std_dev = float(np.std(recent_values))
-        
-        # Define a small improvement threshold based on the verbose (sensitivity) parameter
-        # Lower verbose => more sensitive (higher threshold required), higher verbose => less sensitive (lower threshold)
-        if verbose <= 0:
-            verbose = 1  # avoid division by zero, treat non-positive verbose as most sensitive
-        threshold = 0.01 / verbose  # e.g., verbose=1 -> 0.01 (1% accuracy), verbose=2 -> 0.005, verbose=10 -> 0.001
-        
-        # Check conditions for learning capability
-        # Condition 1: If net improvement is below the threshold (no significant upward trend)
-        if net_improvement <= threshold:
-            return False  # No meaningful overall improvement (plateau or very slow progress)
-        # Condition 2: If accuracy fluctuates a lot without a clear upward trend (improvement overshadowed by noise)
-        if net_improvement > 0 and std_dev > net_improvement:
-            return False  # Variations are larger than the overall improvement (no clear upward trend)
-        
-        # If neither condition triggered, the model is likely still learning (trend is upward)
-        return True
-
-
-    def learning_capable(self, epochsInGeneration=50):
-        """Check if the model is still capable of learning based on accuracy history"""
-        if 'accuracy' not in self.Y or len(self.Y['accuracy']) == 0:
-            print("No accuracy data available")
-            return True
-            
-        # Use all available data points up to patience
-        recent_accuracies = np.array(self.Y['accuracy'][-min(len(self.Y['accuracy']), epochsInGeneration):])
-        return History._check_learning_capability(recent_accuracies)
-
-    def get_last(self, key):
-        return self.Y[key][-1]
-
-    def draw_hist(self, label, path):
-        if not config.SAVE_PLOTS: return
-        for key in self.Y.keys():
-            xc = range(0, len(self.Y[key]))
-            plt.figure()
-            plt.plot(xc, get_list_as_numpy_array(self.Y[key]), label = key)
-            plt.legend()
-            try:
-                plt.savefig(path + "/" + label + "_" + key + ".png")
-            except Exception as e:
-                print(f"Error saving plot: {e}")
-            plt.close()
-            
-    def save(self, path):
-        dict = {}
-        dict['keys'] = {}
-        for key in self.Y.keys():
-            dict['keys'][key] = np.array(get_list_as_numpy_array(self.Y[key]))
-        dict['last_img_id'] = str(self.last_img_id)
-        dict['description'] = str(self.description)
-        dict['best_train_acc'] = str(self.best_train_acc)
-        dict['best_test_acc'] = str(self.best_test_acc)
-        with open(path, 'w+') as f:
-            f.write(json.dumps(dict, cls=NumpyArrayEncoder))
-        with open(path+"_description.txt", 'w') as f:
-            f.write(self.description)
-            
-    def load(self, path):
-        with open(path, "r") as f:
-            data = json.load(f)
-        self.Y = {}
-        self.last_img_id = int(data['last_img_id'])
-        self.description = str(data['description'])
-        if 'best_train_acc' in data.keys():
-            self.best_train_acc = float(data['best_train_acc'])
-            self.best_test_acc = float(data['best_test_acc'])
-        for key in data['keys'].keys():
-            self.Y[key] = list(np.asarray(data['keys'][key]))
 
 
 class Layer_Type(Enum):
@@ -1388,171 +1098,3 @@ def Resize(x, shape):
 
 
 
-class Storage:
-    
-    def __init__(self):
-        pass
-    
-
-    def loadModel(path):
-        with open(path, 'r') as json_file:
-            model_dict = json.load(json_file)
-        return Storage.dictToModel(model_dict)
-
-    def saveModel(model, path):
-        with open(path, 'w') as json_file:
-            json.dump(Storage.modelToDict(model), json_file, indent=4)
-        pass
-    
-    def modelToDict(model):
-        dict_main = {}
-        dict_main['settings'] = {}
-        dict_main['settings']['batch_size'] = model.batch_size
-        dict_main['settings']['loss_function'] = str(model.loss_function.__name__)
-        dict_main['settings']['output_size'] = model.output_size
-        dict_main['settings']['hidden_size'] = model.hidden_size
-        dict_main['settings']['avaible_id'] = model.avaible_id
-        dict_main['settings']['activation_fun'] = str(model.activation_fun.__name__)
-        dict_main['settings']['convolution'] = model.convolution
-        dict_main['settings']['input_size'] = model.input_size
-        dict_main['settings']['input_shape'] = model.input_shape
-        dict_main['settings']['kernel_size'] = model.kernel_size
-        dict_main['settings']['depth'] = model.depth
-        dict_main['settings']['optimizer'] = OptimizerFactory.ToDict(model.optimizer)
-        dict_main['hidden_layers'] = {}
-        dict_main['input_layers'] = {}
-        dict_main['output_layer'] = {}
-        for layer in model.hidden_layers:
-            dict_main['hidden_layers'][layer.id] = Storage.layerToDict(layer)
-        for layer in model.input_layers:
-            dict_main['input_layers'][layer.id] = Storage.layerToDict(layer)
-        dict_main['output_layer'][layer.id] = Storage.layerToDict(model.output_layer)
-        return dict_main
-
-    def layerToDict(layer):
-        dict_main = {}
-        dict_main['id'] = str(layer.id)
-        dict_main['act_fun'] = str(layer.act_fun.__name__)
-        dict_main['is_ending'] = layer.is_ending
-        dict_main['is_starting'] = layer.is_starting
-        dict_main['input_layers_ids'] = [str(i) for i in layer.input_layers_ids] 
-        dict_main['output_layers_ids'] = [str(i) for i in layer.output_layers_ids]
-        dict_main['optimizer'] = OptimizerFactory.ToDict(layer.optimizer)
-        dict_reshapers = {}
-        for resheper_key in layer.reshspers.keys():
-            reshsper_id = str(resheper_key)
-            (resheper_from, resharper_to) = resheper_key
-            dict_reshapers[reshsper_id] = {}
-            dict_reshapers[reshsper_id]['from'] = resheper_from
-            dict_reshapers[reshsper_id]['to'] = resharper_to
-            dict_reshapers[reshsper_id]["matrix"] =  get_numpy_array(layer.reshspers[resheper_key]).tolist()
-        dict_main['reshapers'] = dict_reshapers
-        dict_main['weights'] = {}
-        if type(layer) == Layer:
-            dict_main['neurons'] = layer.neurons
-            dict_main['input_size'] = layer.input_size
-            dict_main['weights']['W'] = get_numpy_array(layer.W).tolist()
-            dict_main['weights']['B'] = get_numpy_array(layer.B).tolist()
-        if type(layer) == Conv:
-            dict_main['weights']['kernels'] = get_numpy_array(layer.kernels).tolist()
-            dict_main['weights']['biases'] = get_numpy_array(layer.biases).tolist()
-            dict_main['conv'] = {}
-            dict_main['conv']['input_shape'] = layer.input_shape
-            dict_main['conv']['input_height'] = layer.input_height
-            dict_main['conv']['input_width'] = layer.input_width
-            dict_main['conv']['input_depth'] = layer.input_depth
-            dict_main['conv']['input_flatten'] = layer.input_flatten
-            dict_main['conv']['depth'] = layer.depth
-            dict_main['conv']['kernel_size'] = layer.kernel_size
-            dict_main['conv']['output_shape'] = layer.output_shape
-            dict_main['conv']['output_flatten'] = layer.output_flatten
-            dict_main['conv']['kernels_shape'] = layer.kernels_shape
-        return dict_main
-        
-
-    def dictToModel(dict_main):
-        batch_size = dict_main['settings']['batch_size']
-        loss_function_name = dict_main['settings']['loss_function']
-        output_size = dict_main['settings']['output_size']
-        hidden_size = dict_main['settings']['hidden_size']
-        avaible_id = dict_main['settings']['avaible_id']
-        activation_fun_name = dict_main['settings']['activation_fun']
-        convolution = dict_main['settings']['convolution']
-        input_shape = dict_main['settings']['input_shape']
-        kernel_size = dict_main['settings']['kernel_size']
-        depth = dict_main['settings']['depth']
-        input_size = dict_main['settings']['input_size']
-        loss_function = Loss.getByName(loss_function_name)
-        activation_fun = Activations.getByName(activation_fun_name)
-        optimizer = OptimizerFactory.FromDict(dict_main['settings']['optimizer'])
-        model = Model(input_size, hidden_size, output_size, loss_function, activation_fun, 1, optimizer)
-        model.batch_size = batch_size
-        model.avaible_id = avaible_id
-        model.convolution = convolution
-        model.kernel_size = kernel_size
-        model.depth = depth
-        model.input_layers = []
-        model.hidden_layers = []
-
-        for layer_id, layer_dict in dict_main['input_layers'].items():
-            layer = Storage.dictToLayer(layer_dict, model)
-            model.input_layers.append(layer)
-
-        for layer_id, layer_dict in dict_main['hidden_layers'].items():
-            layer = Storage.dictToLayer(layer_dict, model)
-            model.hidden_layers.append(layer)
-
-        for layer_id, layer_dict in dict_main['output_layer'].items():
-            layer = Storage.dictToLayer(layer_dict, model)
-            model.output_layer = layer
-        return model
-
-
-    def dictToLayer(layer_dict, model):
-        layer_id = str(layer_dict['id'])
-        act_fun_name = layer_dict['act_fun']
-        is_ending = layer_dict['is_ending']
-        is_starting = layer_dict['is_starting']
-        input_layers_ids = layer_dict['input_layers_ids']
-        output_layers_ids = layer_dict['output_layers_ids']
-        reshapers = layer_dict['reshapers']
-        act_fun = Activations.getByName(act_fun_name)
-        optimizer = OptimizerFactory.FromDict(layer_dict['optimizer'])
-        if 'conv' in layer_dict:
-            conv_dict = layer_dict['conv']
-            layer = Conv(layer_id, model, conv_dict['input_shape'], conv_dict['kernel_size'], conv_dict['depth'], act_fun, optimizer)
-            layer.input_height = conv_dict['input_height']
-            layer.input_width = conv_dict['input_width']
-            layer.input_depth = conv_dict['input_depth']
-            layer.input_flatten = conv_dict['input_flatten']
-            layer.output_shape = conv_dict['output_shape']
-            layer.output_flatten = conv_dict['output_flatten']
-            layer.kernels_shape = conv_dict['kernels_shape']
-        else:
-            input_size = layer_dict['input_size']
-            neurons = layer_dict['neurons']
-            layer = Layer(layer_id, model, input_size, neurons, act_fun, Layer_Type.RANDOM,  optimizer.getDense())
-        
-        if is_ending: layer.set_as_ending()
-        if is_starting : layer.set_as_starting()
-        layer.input_layers_ids = input_layers_ids
-        layer.output_layers_ids = output_layers_ids
-        reshapers_dict = {}
-        for reshaper_id, reshaper_info in reshapers.items():
-            reshaper_from = reshaper_info['from']
-            reshaper_to = reshaper_info['to']
-            reshaper_matrix = np.array(reshaper_info['matrix'])
-            reshapers_dict[(reshaper_from, reshaper_to)] = reshaper_matrix
-        layer.reshspers = reshapers_dict
-
-        if 'weights' in layer_dict:
-            weights_dict = layer_dict['weights']
-            if 'W' in weights_dict:
-                layer.W = np.array(weights_dict['W'])
-            if 'B' in weights_dict:
-                layer.B = np.array(weights_dict['B'])
-            if 'kernels' in weights_dict:
-                layer.kernels = np.array(weights_dict['kernels'])
-            if 'biases' in weights_dict:
-                layer.biases = np.array(weights_dict['biases'])
-        return layer
