+++ growingnn/utils/__init__.py
@@ -1,4 +1,12 @@
 # Utils package for growingnn
 from .stoppers import EmptyStopper, AccuracyStopper, ParameterCountStopper, AccuracyAndReductionStopper
+from .loss import Loss
+from .activations import Activations
+from .lr_scheduler import LearningRateScheduler
+from .history import History
+from .storage import Storage
 
-__all__ = ['EmptyStopper', 'AccuracyStopper', 'ParameterCountStopper', 'AccuracyAndReductionStopper']
+__all__ = [
+    'EmptyStopper', 'AccuracyStopper', 'ParameterCountStopper', 'AccuracyAndReductionStopper',
+    'Loss', 'Activations', 'LearningRateScheduler', 'History', 'Storage'
+]
