+++ tests/test_config_runtime.py
@@ -8,7 +8,8 @@ from testSuite import mode
 import unittest
 import numpy as np
 from growingnn.config import config
-from growingnn.structure import Model, Layer, Activations, Loss
+from growingnn.structure import Model, Layer
+from growingnn.utils import Activations, Loss
 from testSuite import mode
 from testDataGenerator import TestDataGenerator
 
@@ -28,7 +29,7 @@ class TestConfigRuntime(unittest.TestCase):
 
     def test_save_plots_change(self):
         """Test if changing save_plots affects history behavior"""
-        from growingnn.structure import History
+        from growingnn.utils import History
         
         # Create history object
         history = History(['accuracy', 'loss'])
